+++ lib/src/throttle.dart
@@ -3,19 +3,19 @@
 // BSD-style license that can be found in the LICENSE file.
 import 'dart:async';
 
+import 'from_handlers.dart';
+
 /// Creates a StreamTransformer which only emits once per [duration], at the
 /// beginning of the period.
 StreamTransformer<T, T> throttle<T>(Duration duration) {
   Timer timer;
 
-  return new StreamTransformer.fromHandlers(handleData: (data, sink) {
+  return fromHandlers(handleData: (data, sink) {
     if (timer == null) {
       sink.add(data);
       timer = new Timer(duration, () {
         timer = null;
       });
     }
-  }, handleDone: (sink) {
-    sink.close();
   });
 }