+++ test/switch_test.dart
@@ -136,7 +136,14 @@ void main() {
       await new Future(() {});
       outer.add([4, 5, 6]);
       await new Future(() {});
-      expect(values, [1, 2, 3, 4, 5, 6]);
+    });
+
+    test('can create a broadcast stream', () async {
+      var outer = new StreamController.broadcast();
+
+      var transformed = outer.stream.transform(switchMap(null));
+
+      expect(transformed.isBroadcast, true);
     });
   });
 }